move file scan outside intel services
