+++ examples/file_scan/README.md
@@ -1,0 +1,11 @@
+# Pangea FileScan Examples
+
+## Setup
+
+Set up environment variables ([Instructions](https://pangea.cloud/docs/getting-started/integrate/#set-environment-variables)) `PANGEA_FILE_SCAN_TOKEN` and `PANGEA_DOMAIN` with your project token configured on Pangea User Console (token should have access to FileScan services [Instructions](https://pangea.cloud/docs/getting-started/configure-services/#configure-a-pangea-service)) and with your pangea domain.
+
+Run examples
+
+```
+% node file_scan_async.mjs
+```
