+++ examples/file_scan/file_scan_sync.mjs
@@ -3,7 +3,7 @@
 import { PangeaConfig, FileScanService, PangeaErrors } from "pangea-node-sdk";
 
 const domain = process.env.PANGEA_DOMAIN;
-const token = process.env.PANGEA_INTEL_TOKEN;
+const token = process.env.PANGEA_FILE_SCAN_TOKEN;
 
 // To work in sync it's need to set up queuedRetryEnabled to true and set up a proper timeout
 // If timeout it's so little service won't end up and will return an AcceptedRequestException anyway
@@ -12,16 +12,16 @@ const config = new PangeaConfig({
   queuedRetryEnabled: true,
   pollResultTimeoutMs: 60 * 1000,
 });
-const intel = new FileScanService(String(token), config);
+const client = new FileScanService(String(token), config);
 
-const yourFilepath = "./intel/file_scan/testfile.pdf";
+const yourFilepath = "./testfile.pdf";
 
 (async () => {
   console.log("Checking file...");
 
   try {
     const request = { verbose: true, raw: true, provider: "crowdstrike" };
-    const response = await intel.fileScan(request, yourFilepath);
+    const response = await client.fileScan(request, yourFilepath);
     console.log("Result:", response.result);
   } catch (e) {
     if (e instanceof PangeaErrors.APIError) {
