+++ examples/file_scan/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "intel_examples",
+  "version": "2.0.0",
+  "description": "Pangea Intel Services examples",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Pangea",
+  "license": "ISC",
+  "dependencies": {
+    "pangea-node-sdk": "^2.2.0"
+  }
+}
