+++ packages/pangea-node-sdk/src/services/index.ts
@@ -9,9 +9,9 @@ import {
   IPIntelService,
   URLIntelService,
   UserIntelService,
-  FileScanService,
 } from "./intel.js";
 import VaultService from "./vault.js";
+import { FileScanService } from "./file_scan.js";
 
 export default {
   AuditService,
